requirements.txt
# Web
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0

# Database
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# ML
xgboost==2.0.3
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
joblib==1.4.2

# LangGraph / LangChain
langgraph==0.0.66
langchain==0.2.1
langchain-groq==0.1.3
langchain-community==0.2.1

# RAG
chromadb==0.5.0
sentence-transformers==2.7.0

# Validation
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Utilities
rich==13.7.1
typer==0.12.3
